[Backend] fix: security

+++ int222_backend/src/main/java/sit/int221/services/SubscriptionService.java
@@ -12,12 +12,9 @@ import sit.int221.repositories.SubscriptionRepository;
 import sit.int221.repositories.UnsubscribeTokenRepository;
 import sit.int221.utils.AnnouncementDisplay;
 
-import java.security.SecureRandom;
 import java.time.Duration;
-import java.time.LocalDateTime;
 import java.time.ZoneId;
 import java.time.ZonedDateTime;
-import java.util.Base64;
 import java.util.List;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.Executors;
@@ -45,10 +42,14 @@ public class SubscriptionService {
         }
     }
 
-    public void unsubscribe(String email, Integer categoryId){
-        Subscription subscription = subscriptionRepository.findByEmailAndCategoryId(email, categoryId);
-        if (subscription != null) {
-            subscriptionRepository.delete(subscription);
+    @Transactional
+    public void unsubscribe(String token, Integer categoryId) {
+        String email = verifyUnsubscribeToken(token);
+        if (email != null) {
+            Subscription subscription = subscriptionRepository.findByEmailAndCategoryId(email, categoryId);
+            if (subscription != null) {
+                subscriptionRepository.delete(subscription);
+            }
         }
     }
 
@@ -83,40 +84,14 @@ public class SubscriptionService {
         CompletableFuture.completedFuture(null);
     }
 
-    public String generateUnsubscribeToken(String email) {
-        UnsubscribeToken existingToken = unsubscribeTokenRepository.findByEmail(email);
-        if (existingToken != null) {
-            unsubscribeTokenRepository.delete(existingToken);
-        }
-
-        SecureRandom secureRandom = new SecureRandom();
-        byte[] randomBytes = new byte[24];
-        secureRandom.nextBytes(randomBytes);
-        String token = Base64.getUrlEncoder().withoutPadding().encodeToString(randomBytes);
-
-        UnsubscribeToken unsubscribeToken = new UnsubscribeToken();
-        unsubscribeToken.setEmail(email);
-        unsubscribeToken.setToken(token);
-        unsubscribeToken.setGeneratedAt(LocalDateTime.now());
-
-        unsubscribeTokenRepository.save(unsubscribeToken);
-
-        return token;
-    }
-
     @Transactional
     public String verifyUnsubscribeToken(String token) {
         UnsubscribeToken unsubscribeToken = unsubscribeTokenRepository.findByToken(token);
-        if (unsubscribeToken != null && !isUnsubscribeTokenExpired(unsubscribeToken.getGeneratedAt())) {
+        if (unsubscribeToken != null) {
             unsubscribeTokenRepository.deleteByToken(token);
             return unsubscribeToken.getEmail();
         } else {
             return null;
         }
-    }
-
-    private boolean isUnsubscribeTokenExpired(LocalDateTime tokenGeneratedAt) {
-        LocalDateTime now = LocalDateTime.now();
-        return tokenGeneratedAt.isBefore(now.minusDays(30));
     }
 }
